store: move download tests to download_test.go
